@@ -1,4 +1,4 @@
-interface TokenResponse {
+export interface TokenResponse {
   token: string,
   expiration: number,
   refresh_token: string
