@@ -87,7 +87,7 @@ namespace SkillersPokerSystem.Controllers
                 var t = CreateAccessToken(user.Id, rt.Value);
                 return Json(t);
             }
-            catch (Exception ex)
+            catch (Exception)
             {
                 return new UnauthorizedResult();
             }
@@ -136,8 +136,9 @@ namespace SkillersPokerSystem.Controllers
                 // ... and send it to the client
                 return Json(response);
             }
-            catch (Exception ex)
+            catch (Exception)
             {
+                
                 return new UnauthorizedResult();
             }
         }
